+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
-## Unpublished
+## [x.Y.0] - YYYY-MM-DD
 
 ### 🛠 Breaking changes
 
 ### 🎉 New features
+- Added support for route color customization
+  - `routeColor`: Customize main route color
+  - `routeAlternateColor`: Customize alternate routes color
+  - `routeCasingColor`: Customize main route casing color
+  - `routeAlternateCasingColor`: Customize alternate routes casing color
+  - `traversedRouteColor`: Customize traversed route color
+  - `maneuverArrowColor`: Customize maneuver arrow color
 
 ### 🐛 Bug fixes